Enhancements: Update UI + Functionality
Enhancements: Improved Block Editor date selection tool. Now requires a checkbox to be selected for this to show/hide. This should simplify the block editor for users.

Added in a shortcode attribute "use_dates" which can stay hidden, takes true/false. This used if the select date option is set inside the Block Editor. Not needed for shortcode.

ENHANCEMENT: Added maximum amount of keywords allowed for a block (3).

Pushed build index.js and index.js.map as well.